+++ dffml/cli/dataflow.py
@@ -73,7 +73,7 @@ class CreateConfig:
     not_linked: bool = field(
         "Do not export dataflows as linked", default=False,
     )
-    seed: List[str] = field(
+    inputs: List[str] = field(
         "Inputs to be added to every context",
         action=ParseInputsAction,
         default_factory=lambda: [],
@@ -92,19 +92,27 @@ class Create(CMD):
 
     async def run(self):
         operations = []
+        operation_names = []
         for load_operation in self.operations:
-            if ":" in load_operation:
+            if "=" in load_operation:
+                opname, operation = load_operation.split("=")
+            else:
+                opname = operation = load_operation
+            operation_names.append(opname)
+            if ":" in operation:
                 operations.extend(
                     map(
                         OperationImplementation._imp,
-                        load(load_operation, relative=True),
+                        load(operation, relative=True),
                     )
                 )
             else:
-                operations += [Operation.load(load_operation)]
+                operations += [Operation.load(operation)]
         async with self.configloader(BaseConfig()) as configloader:
             async with configloader() as loader:
-                dataflow = DataFlow.auto(*operations)
+                dataflow = DataFlow(
+                    operations=dict(zip(operation_names, operations))
+                )
                 # flow argument key is of the form opname.inputs/conditions.keyname
                 for v, k in self.flow:
                     *opname, val_type, key = split_dot_seperated(k)
@@ -116,15 +124,27 @@ class Create(CMD):
                 for v, k in self.config:
                     traverse_set(dataflow.configs, k, value=v)
                 exported = dataflow.export(linked=not self.not_linked)
-                if self.seed:
+                if self.inputs:
                     if not "seed" in exported:
                         exported["seed"] = []
-                    exported["seed"].extend(
-                        [
-                            {"value": val, "definition": def_name}
-                            for val, def_name in self.seed
-                        ]
-                    )
+                    for val, input_info in self.inputs:
+                        if "=" in input_info:
+                            def_name, origin = input_info.split(
+                                "=", maxsplit=1
+                            )
+                            # self.inputs is of the form val=def_name=origin
+                            exported["seed"].append(
+                                {
+                                    "value": val,
+                                    "definition": def_name,
+                                    "origin": origin,
+                                }
+                            )
+                        else:
+                            # self.inputs is of the form val=def_name
+                            exported["seed"].append(
+                                {"value": val, "definition": input_info}
+                            )
                 print((await loader.dumpb(exported)).decode())