+++ dffml/plugins.py
@@ -2,7 +2,7 @@ import sys
 
 CORE_PLUGINS = [
     ("configloader", "yaml"),
-    ("configloader", "png"),
+    ("configloader", "image"),
     ("model", "scratch"),
     ("model", "scikit"),
     ("model", "tensorflow"),
@@ -16,6 +16,7 @@ CORE_PLUGINS = [
     ("operations", "binsec"),
     ("operations", "deploy"),
     ("operations", "image"),
+    ("operations", "nlp"),
     ("service", "http"),
     ("source", "mysql"),
 ]