+++ examples/MNIST/create_dataflow.sh
@@ -1,4 +1,4 @@
 dffml dataflow create multiply associate_definition -configloader yaml \
-    -seed '{"image": "product"}'=associate_spec 0.00392156862=multiplier_def \
+    -inputs '{"image": "product"}'=associate_spec 0.00392156862=multiplier_def \
     -flow '[{"seed": ["image"]}]'=multiply.inputs.multiplicand |
     tee normalize.yaml