+++ model/tensorflow_hub/tests/test_model.py
@@ -62,7 +62,7 @@ class TestTextClassificationModel(AsyncTestCase):
         async with self.sources as sources, self.model as model:
             target_name = model.config.predict.name
             async with sources() as sctx, model() as mctx:
-                async for record in mctx.predict(sctx.records()):
+                async for record in mctx.predict(sctx):
                     prediction = record.prediction(target_name).value
                     self.assertIn(prediction, ["0", "1"])