+++ model/transformers/dffml_model_transformers/ner/ner_model.py
@@ -24,10 +24,10 @@ from transformers import (
 from dffml import export
 from dffml.record import Record
 from dffml.base import config, field
-from dffml.source.source import Sources
 from dffml.feature.feature import Feature
 from dffml.model.accuracy import Accuracy
 from dffml.util.entrypoint import entrypoint
+from dffml.source.source import Sources, SourcesContext
 from dffml.model.model import ModelContext, Model, ModelNotTrained
 
 from .utils import (
@@ -409,7 +409,7 @@ class NERModelContext(ModelContext):
         return Accuracy(result["eval_f1"])
 
     async def predict(
-        self, records: AsyncIterator[Record]
+        self, sources: SourcesContext
     ) -> AsyncIterator[Tuple[Record, Any, float]]:
         if not os.path.isfile(
             os.path.join(self.parent.config.output_dir, "tf_model.h5")
@@ -422,7 +422,9 @@ class NERModelContext(ModelContext):
                 cache_dir=self.parent.config.cache_dir,
             )
 
-        async for record in records:
+        async for record in sources.with_features(
+            [self.parent.config.words.name]
+        ):
             sentence = record.features([self.parent.config.words.name])
             df = self.pd.DataFrame(sentence, index=[0])
             test_dataset = self.get_dataset(df, self.tokenizer, mode="test",)